Add option to customize loading icon
With project being customized with different look and feels, the loading icon should be able to
be customized as well. The option that gives the developer the most creative freedom is to design
their theme specific loader and have catalin just show or hide it based on the id. The default
for the value is 'loader,' the icon bundled with Catalin.